+++ src/App/App.jsx
@@ -8,8 +8,21 @@ import {
   Error,
 } from '../Pages/index'
 import Navbar from '../Components/Navbar'
+import { useEffect } from 'react'
+import { useDispatch, useSelector } from 'react-redux'
+import { selectCoins } from '../Pages/Main/MainSlice'
+import { getCurrencies } from '../Pages/Main/MainSlice'
 
 export default function App() {
+  const dispatch = useDispatch()
+  const coins = useSelector(selectCoins)
+
+  useEffect(() => {
+    if (coins.length === 0) {
+      dispatch(getCurrencies())
+    }
+  }, [dispatch, coins.length])
+
   return (
     <div className="app">
       <Routes>
